remove SettingsActivity
