+++ app/src/main/AndroidManifest.xml
@@ -86,10 +86,6 @@
         <category android:name="android.intent.category.LAUNCHER" />
       </intent-filter>
     </activity>
-
-    <activity
-        android:name="com.asfoundation.wallet.ui.settings.SettingsActivity"
-        android:label="" />
     <activity
         android:name="com.asfoundation.wallet.ui.balance.TransactionDetailActivity"
         android:label="" />
