--- app/src/main/java/com/asfoundation/wallet/ui/settings/SettingsActivityView.kt
@@ -1,8 +1,0 @@
-package com.asfoundation.wallet.ui.settings
-
-import io.reactivex.Observable
-
-interface SettingsActivityView {
-
-  fun authenticationResult(): Observable<Boolean>
-}
